refactor: 타입 선언 분리
